Make the max commit message subject and line lengths configurable
Instead of hard-coding the maximum commit message subject and line
lengths to 65 and 70 characters respectively, make them user
configurable.

The user can add a new config file "commitmessage.config" in the site's
etc folder, which specifies the maximum subject and line lengths.

If the config file does not exist, or is empty, the default values (65
and 70) are used.

Change-Id: I43dacc2dd281fd521a156b19ec145fe299f658ef