Allow to reject commits when subject or line length exceeds limit
Currently the commit message validator just emits a warning if
the commit message subject or line length exceeds the configured
limit.

Add a new configuration item that will allow the user to specify
that a commit should be rejected with an error if the commit message
subject or line length is too long.

Change-Id: I9a1355eef356fa25ee978607015aa664bf6119a0