Update patch set 2
Patch Set 2:

This is rebased on the new initial version.

To address Luca's comment, I just had a look at how the project config in refs/meta/config works.  It looks like it will need support in core gerrit, and I think that kind of defeats the point of moving this functionality out to a plugin.

Maybe I'm wrong though.  Is it possible to put such config in refs/meta/config without having to add surrounding framework code?

Patch-set: 2