Validation for commit message body
Add a validation to check that the commit message contains a
body or description of the change.  Provide an option to allow
gerrit to reject the change if a description is missing from
the commit message.

Do not enforce commits from Inline edit's 'edit project'
feature because it stages the commit without a commit
message body.

Change-Id: I79b27e6bf9404518aefefcf057946fb477ca489d