Validation for commit message body
Add a validation to check that the commit message contains a
body or description of the change.  Provide an option to allow
gerrit to reject the change if a description is missing from
the commit message.

This change depends on a core gerrit change (I1e07e7) to allow
it to work with inline edit's 'edit project' feature.

Change-Id: I79b27e6bf9404518aefefcf057946fb477ca489d