Update patch set 2
Patch Set 2:

Changed to WIP because I just noticed that setting this to reject a commit will break inline edit.  I believe it's because inline edit does not add a description when it pushes to a temporary ref location.  Thus an inline edit cannot be 'staged' for editing.

Patch-set: 2
Label: Verified=0